marketing/requirements.txt
streamlit>=1.28.0
crewai
crewai-tools
python-dotenv
pathlib 